Moved post assessment functionality to celery task (#237)
