+++ openedx/features/pakx/lms/overrides/tasks.py
@@ -22,6 +22,7 @@ from openedx.core.djangoapps.user_api.preferences.api import get_user_preference
 from openedx.core.lib.celery.task_utils import emulate_http_request
 from openedx.features.pakx.lms.overrides.message_types import ContactUs, CourseProgress
 from openedx.features.pakx.lms.overrides.models import CourseProgressStats
+from openedx.features.pakx.lms.overrides.post_assessment import check_and_unlock_user_milestone
 from openedx.features.pakx.lms.overrides.utils import (
     create_dummy_request,
     get_course_progress_percentage,
@@ -188,3 +189,13 @@ def update_course_progress_stats():
                 item.email_reminder_status = CourseProgressStats.REMINDER_SENT
                 fields_list.append('email_reminder_status')
         item.save(update_fields=fields_list)
+
+
+@task(name='unlock_subsections')
+def unlock_subsections():
+    """Start checking and unlocking milestones"""
+
+    progress_models = CourseProgressStats.objects.filter(progress__lt=100).select_related('enrollment')
+    log.info("Fetching records, found {} active models".format(len(progress_models)))
+    for item in progress_models:
+        check_and_unlock_user_milestone(item.enrollment.user, text_type(item.enrollment.course_id))
